kememtembem/requirements.txt
base58==2.1.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colander==2.0
ecdsa==0.18.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
IneryPy==0.1
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
ordereddict==1.1
python-dotenv==1.0.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
translationstring==1.4
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0